Extract S3DeleteFiles
